PagedListAdapter 생성, id 값 Int > Long으로 변경

+++ app/src/main/java/com/ddd/airplane/repository/network/service/ChatService.kt
@@ -19,7 +19,7 @@ interface ChatService {
      */
     @GET(ServerUrl.GET_ROOM_INFO)
     fun getRoom(
-        @Path("roomId") roomId: Int
+        @Path("roomId") roomId: Long
     ): Single<ChatRoomData>
 
     /**
@@ -27,7 +27,7 @@ interface ChatService {
      */
     @GET(ServerUrl.GET_ROOM_MESSAGE)
     fun getRoomMessages(
-        @Path("roomId") roomId: Int,
+        @Path("roomId") roomId: Long,
         @Query("baseMessageId") baseMessageId: Int,
         @Query("size") size: Int,
         @Query("direction") direction: String
